requirements.txt

pyyaml==6.0.1
python-dateutil==2.9.0
pathlib2==2.3.7
